% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylog.R
\name{mylog}
\alias{mylog}
\title{Evaluates the natural logarithm}
\usage{
mylog(x)
}
\arguments{
\item{x}{A number or vector of numbers.}
}
\value{
The logarithm of every element in \code{x}
}
\description{
Evaluates the natural logarithm
}
\examples{
mylog(c(1, 2, 3, 4, 5))
mylog(100)
}
